+++ CHANGELOG.md
@@ -41,6 +41,7 @@ All notable changes to this project will be documented in this file.
 - [#2447] SVG import - gradients can inherit (href) from other gradient types (radial vs linear)
 - [#2450] Morphshape replace button/menu is not working (throws exception due to missing icon)
 - [#2355] AS1/2 Simplify expressions feature colliding with some other features like hex values
+- Exception on FFDec start when simple editor is on
 
 ## [22.0.2] - 2025-01-17
 ### Added